+++ cmd/frontend/graphqlbackend/search_results.go
@@ -22,6 +22,7 @@ import (
 	"gopkg.in/inconshreveable/log15.v2"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/goroutine"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/inventory/filelang"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/pkg/search"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/pkg/search/query"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
@@ -230,14 +231,31 @@ func (sr *searchResultsResolver) DynamicFilters() []*searchFilterResolver {
 		repoToMatchCount[uri] += lineMatchCount
 		add(filter, uri, repoToMatchCount[uri], limitHit, "repo")
 	}
-	addFileFilter := func(filematchPath string, lineMatchCount int, limitHit bool) {
-		if ext := path.Ext(filematchPath); ext != "" {
-			value := fmt.Sprintf(`file:%s$`, regexp.QuoteMeta(ext))
-			add(value, value, lineMatchCount, false, "file")
-		}
+
+	addFileFilter := func(fileMatchPath string, lineMatchCount int, limitHit bool) {
 		for _, ff := range commonFileFilters {
-			if ff.Regexp.MatchString(filematchPath) {
+			if ff.Regexp.MatchString(fileMatchPath) {
 				add(ff.Filter, ff.Filter, lineMatchCount, limitHit, "file")
+			}
+		}
+	}
+
+	addLangFilter := func(fileMatchPath string, lineMatchCount int, limitHit bool) {
+		extensionToLanguageLookup := func(ext string) string {
+			for _, lang := range filelang.Langs {
+				for _, langExt := range lang.Extensions {
+					if ext == langExt {
+						return strings.ToLower(lang.Name)
+					}
+				}
+			}
+			return ""
+		}
+		if ext := path.Ext(fileMatchPath); ext != "" {
+			language := extensionToLanguageLookup(path.Ext(fileMatchPath))
+			if language != "" {
+				value := fmt.Sprintf(`lang:%s`, language)
+				add(value, value, lineMatchCount, limitHit, "lang")
 			}
 		}
 	}
@@ -249,6 +267,7 @@ func (sr *searchResultsResolver) DynamicFilters() []*searchFilterResolver {
 				rev = *result.fileMatch.inputRev
 			}
 			addRepoFilter(string(result.fileMatch.repo.Name), rev, len(result.fileMatch.LineMatches()))
+			addLangFilter(result.fileMatch.JPath, len(result.fileMatch.LineMatches()), result.fileMatch.JLimitHit)
 			addFileFilter(result.fileMatch.JPath, len(result.fileMatch.LineMatches()), result.fileMatch.JLimitHit)
 
 			if len(result.fileMatch.symbols) > 0 {
@@ -305,7 +324,7 @@ type searchFilterResolver struct {
 	// whether the results returned for a repository are incomplete
 	limitHit bool
 
-	// the kind of filter. Should be "repo" or "file".
+	// the kind of filter. Should be "repo", "file", or "lang".
 	kind string
 
 	// score is used to select potential filters